Separate ops into sub-folder and move runtime ops to func calls pass earlier
